Backend design

+++ hknweb/tutoring/views.py
@@ -1,9 +1,10 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required, permission_required
 from django.contrib import messages
+from django.forms import formset_factory
 from .models import Slot, Tutor, Course
-from .forms import ClassPreferenceForm
-
+from .forms import ClassPreferenceForm 
+import json
 
 def index(request):
     days = [name for _, name in Slot.DAY_CHOICES]
@@ -27,11 +28,16 @@ def index(request):
     print(context)
     return render(request, 'tutoring/index.html', context)
 
+def tutor_slot_preference(request):
+    if request.method == 'POST':
+        #process form
+        return
+    SlotPreferenceFormSet = formset_factory()
 # @permission_required('events.add_event', login_url='/accounts/login/')
 @login_required(login_url='/accounts/login/')
-def tutor_class_preference(request):
-    form = ClassPreferenceForm(request.POST or None)
+def tutor_course_preference(request):
     if request.method == 'POST':
+        form = ClassPreferenceForm(request.POST or None)
         if form.is_valid():
             courses = form.cleaned_data.get("courses")
             if Tutor.objects.filter(user = request.user).exists():
@@ -51,12 +57,28 @@ def tutor_class_preference(request):
     return render(request, 'tutoring/preferences.html', {'form': ClassPreferenceForm(None)})
 
 
+# @permission_required('events.add_event', login_url='/accounts/login/')
+@login_required(login_url='/accounts/login/')
+def tutor_slot_preference(request):
+    return
+
+def prepare_algorithm_input(request):
+    if request.method == 'GET':
+        
+
 def generate_schedule(request):
-    Slot.objects.all().delete()
-    for room, _ in Slot.ROOM_CHOICES:
-        for day, _ in Slot.DAY_CHOICES:
-            for hour, _ in Slot.HOUR_CHOICES:
-                print(room, day, hour)
-                slot = Slot(hour = hour, day = day, room = room)
-                slot.save()
+    if request.method == 'POST':
+        json_data = json.loads(request.body)
+        slot_id = 121
+        Slot.objects.all().delete()
+        for hour, _ in Slot.HOUR_CHOICES:
+            for day, _ in Slot.DAY_CHOICES:
+                for office, _ in Slot.ROOM_CHOICES:
+                    slot = Slot(hour = hour, day = day, room = room, slot_id = id)
+                    tutor_ids = json_data[slot_id]
+                    for tutor_id in tutor_ids:
+                        tutor = Tutor.objects.get(id=tutor_id)
+                        slot.tutors.add(tutor)
+                    slot.save()
+                    slot_id += 1
     return index(request)
